+++ src/EntityGraphQL/Schema/IMethodType.cs
@@ -5,5 +5,6 @@ namespace EntityGraphQL.Schema
     public interface IMethodType
     {
         Type GetArgumentType(string argName);
+        bool HasArgumentType(string argName);
     }
 }